テーブル主キーの生成方法を変更する

+++ backend/src/main/java/com/example/stamp_app/controller/param/MicroControllerPostParam.java
@@ -6,14 +6,12 @@ import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class MicroControllerPostParam {
 
     @NotNull
     @Positive
-    private BigInteger userId;
+    private Long userId;
 
     @NotBlank
     @Pattern(regexp = "^[0-9a-fA-F]{2}(:[0-9a-fA-F]{2}){5}$")
