+++ backend/src/main/java/com/example/stamp_app/repository/AccountRepository.java
@@ -3,7 +3,6 @@ package com.example.stamp_app.repository;
 import com.example.stamp_app.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 import java.util.UUID;
 
 public interface AccountRepository extends JpaRepository<Account, UUID> {
@@ -11,5 +10,5 @@ public interface AccountRepository extends JpaRepository<Account, UUID> {
 
     Account findByUuid(UUID uuid);
 
-    Account findById(BigInteger id);
+    Account findById(Long id);
 }
